backend/requirements.txt
aiosqlite
click
fastapi
pydantic<2.0.0
tortoise-orm
typing-extensions
uvicorn
